requirements.txt
Django>=4.2
pytz
sqlparse
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
gunicorn==20.1.0
whitenoise==6.4.0
Pillow
dj-database-url
psycopg2-binary
